Temporary Second Version of Final Frame
I added another folder with basically a copy of the finalized frame but I am altering this to include the 1/8" thick L angles if thats what we end up using. In the event that we use this instead, I'll dissolve the original 'Final Frame 2025' and replace it with this model and this folder.